Added architecture-specific compilation unit to simplify porting and hide implementation details
