+++ src/toneAC.h
@@ -90,31 +90,9 @@ to control a two-pin dual LED).
     #include <WProgram.h>
   #endif
 
-  #if defined (__AVR_ATmega32U4__) || defined(__AVR_ATmega640__) || defined(__AVR_ATmega1280__) || defined(__AVR_ATmega1281__) || defined(__AVR_ATmega2560__) || defined(__AVR_ATmega2561__) || defined(__AVR_AT90USB646__) || defined(__AVR_AT90USB1286__)
-    #define PWMT1AMASK DDB5
-    #define PWMT1BMASK DDB6
-    #define PWMT1DREG DDRB
-    #define PWMT1PORT PORTB
-  #elif defined(__AVR_ATmega1284P__) || defined(__AVR_ATmega644__) || defined(__AVR_ATmega644P__)
-    #define PWMT1AMASK DDD4
-    #define PWMT1BMASK DDD5
-    #define PWMT1DREG DDRD
-    #define PWMT1PORT PORTD
-  #else
-    #define PWMT1AMASK DDB1
-    #define PWMT1BMASK DDB2
-    #define PWMT1DREG DDRB
-    #define PWMT1PORT PORTB
-  #endif
-
-  #if defined(__AVR_ATmega8__) || defined(__AVR_ATmega128__)
-    #define TIMSK1 TIMSK
-  #endif
-
   #define NOTONEAC 0
   #define PLAY_FOREVER 0
 
   void toneAC(unsigned long frequency = NOTONEAC, uint8_t volume = 10, unsigned long length = PLAY_FOREVER, uint8_t background = false);
-  void toneAC_playNote(unsigned long frequency, uint8_t volume);
   void noToneAC();
 #endif
